modernize LhcTrackAnalyzer
