--- Alignment/HIPAlignmentAlgorithm/interface/LhcTrackAnalyzer.h
@@ -1,101 +1,0 @@
-#ifndef LhcTrackAnalyzer_h
-#define LhcTrackAnalyzer_h
-
-#include "FWCore/Framework/interface/Frameworkfwd.h"
-#include "FWCore/Framework/interface/one/EDAnalyzer.h"
-#include "FWCore/Framework/interface/Event.h"
-#include "FWCore/Framework/interface/EventSetup.h"
-#include "FWCore/Framework/interface/MakerMacros.h"
-#include "FWCore/ParameterSet/interface/ParameterSet.h"
-
-#include "TFile.h"
-#include "TH1D.h"
-#include "TH1I.h"
-#include "TH2D.h"
-#include "TTree.h"
-
-#include "DataFormats/TrackerRecHit2D/interface/SiStripRecHit2D.h"
-
-#include "DataFormats/TrackCandidate/interface/TrackCandidate.h"
-#include "DataFormats/TrackCandidate/interface/TrackCandidateCollection.h"
-#include "DataFormats/TrackReco/interface/Track.h"
-#include "DataFormats/TrackReco/interface/TrackFwd.h"
-#include "DataFormats/TrackingRecHit/interface/TrackingRecHit.h"
-#include "DataFormats/DetId/interface/DetId.h"
-#include "DataFormats/SiStripDetId/interface/StripSubdetector.h"
-#include "DataFormats/SiPixelDetId/interface/PixelSubdetector.h"
-
-//FOR CLUSTERINFO
-#include "DataFormats/TrackerRecHit2D/interface/SiStripRecHit2D.h"
-#include "DataFormats/SiStripDigi/interface/SiStripDigi.h"
-#include "DataFormats/SiStripCluster/interface/SiStripCluster.h"
-//#include "DataFormats/SiStripCluster/interface/SiStripClusterInfo.h"
-#include "DataFormats/Common/interface/DetSetVector.h"
-
-#include "Geometry/Records/interface/TrackerDigiGeometryRecord.h"
-#include "Geometry/TrackerGeometryBuilder/interface/TrackerGeometry.h"
-#include "DataFormats/GeometryVector/interface/GlobalPoint.h"
-#include "DataFormats/GeometryVector/interface/LocalPoint.h"
-#include "Geometry/CommonDetUnit/interface/GeomDet.h"
-
-// system include files
-#include <iostream>
-#include <memory>
-#include <string>
-#include <sstream>
-#include <vector>
-#include <map>
-
-//
-// class decleration
-//
-
-class LhcTrackAnalyzer : public edm::one::EDAnalyzer<> {
-public:
-  explicit LhcTrackAnalyzer(const edm::ParameterSet&);
-  ~LhcTrackAnalyzer() override;
-
-private:
-  void beginJob() override;
-  void analyze(const edm::Event&, const edm::EventSetup&) override;
-  void endJob() override;
-
-  // ----------member data ---------------------------
-  edm::InputTag TrackCollectionTag_;
-  edm::InputTag PVtxCollectionTag_;
-  bool debug_;
-
-  // Output
-  std::string filename_;
-  TFile* rootFile_;
-  TTree* rootTree_;
-
-  // Root-Tuple variables :
-  //=======================
-  void SetVarToZero();
-
-  static const int nMaxtracks_ = 3000;
-  int nTracks_;
-  int run_;
-  int event_;
-  double pt_[nMaxtracks_];
-  double eta_[nMaxtracks_];
-  double phi_[nMaxtracks_];
-  double chi2_[nMaxtracks_];
-  double chi2ndof_[nMaxtracks_];
-  int charge_[nMaxtracks_];
-  double qoverp_[nMaxtracks_];
-  double dz_[nMaxtracks_];
-  double dxy_[nMaxtracks_];
-  double xPCA_[nMaxtracks_];
-  double yPCA_[nMaxtracks_];
-  double zPCA_[nMaxtracks_];
-  int trkAlgo_[nMaxtracks_];
-  int trkQuality_[nMaxtracks_];
-  int isHighPurity_[nMaxtracks_];
-  int validhits_[nMaxtracks_][7];
-  bool goodbx_;
-  bool goodvtx_;
-};
-
-#endif
